requirements.txt
cassandra-driver==3.25.0
python-dotenv==0.21.0 
pyspark==3.3.0